requirements.txt
streamlit
tensorflow-cpu
pillow
opencv-python-headless
numpy